Added style files for bookDetail component
