+++ app/screens/Lists/BooksList/components/BookDetail/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    alignSelf: 'stretch',
+    justifyContent: 'flex-start',
+    flex: 1
+  },
+  image: {
+    alignSelf: 'stretch',
+    height: 200
+  },
+  text: {
+    fontSize: 18,
+    marginTop: 10,
+  }
+});
